UnitTest Overhaul (#115)
* UnitTestRunner

* UnitTest helper classes

* Converted AddUT to new UT framework

* Converted AndUT

* Convert BicUT

* Convert Cmn and Cmp

* Converted remaining arith and logic UTs

* Converted FlowCtrl UTs

* Combined CMake files

* Unit Testing Framework Documentation

* Convert hashmap and stack

* Convert listUT

* Switch to use Catch2

* Add catch install

* Try this

* Maybe?

* Switch to /

* sdfsd

* cd

* cd ..

* add ls

* Change to catch2

* Removed vcpkg

* Fail test

* added always

* Converted linked list tests

* Convert remaining data structures

* Converted AND and ADD

* Converted BicUT

* CMN, CMP, EOR, and MOV conversion

* LSL and LSR conversion

* Update Master CI

* Converted last of ArithAndLogic

* Convert FlowCtrl suite

* Convert LDR and STR UT

* Small fixes

* Last of the conversions

+++ src/Instructions/MultipleMemoryInstruction.cpp
@@ -32,7 +32,7 @@ void MultipleMemoryInstruction::Execute(Registers& rProcessRegisters)
     }
 
     // For every register in the list...
-    for (int i = 0; i < m_registerList.GetLength(); i++)
+    for (uint8_t i = 0; i < m_registerList.GetLength(); i++)
     {
         // Handle IB and DB
         if (m_addressingMode == AddressingMode::IB)
